+++ src/lib/results/index.ts
@@ -1,4 +1,5 @@
 export { setResult } from "./set-result";
+export { unsetResult } from "./unset-result";
 export { getResultsByCeremony, getResultByCategory } from "./get-results";
 export {
   checkResultsPermission,